requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic[email]==2.5.3
flask==3.0.0
requests==2.31.0
